email_pipeline/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
asyncio==3.4.3
charset-normalizer==3.4.2
click==8.2.1
dataclasses==0.6
deprecation==2.1.0
beautifulsoup4==4.12.3
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.115.12
frozenlist==1.6.0
gotrue==2.12.0
greenlet==3.2.2
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
jiter==0.9.0
joblib==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-core==0.3.63
langchain-openai==0.3.18
langchain-text-splitters==0.3.8
langgraph==0.4.8
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langsmith==0.3.44
multidict==6.4.3
numpy==2.2.5
openai==1.77.0
orjson==3.10.18
ormsgpack==1.10.0
pandas==2.2.3
pluggy==1.6.0
postgrest==1.0.1
propcache==0.3.1
psycopg2-binary==2.9.10
pydantic==2.11.5
pydantic-core==2.33.2
pyjwt==2.10.1
pytest==8.3.5
pytest-mock==3.14.0
python-dotenv==1.1.0
pytz==2025.2
realtime==2.4.3
regex==2024.11.6
requests-toolbelt==1.0.0
scikit-learn==1.6.1
scipy==1.15.2
sniffio==1.3.1
sqlalchemy==2.0.41
starlette==0.46.2
storage3==0.11.3
strenum==0.4.15
supabase==2.15.1
supafunc==0.9.4
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tqdm==4.67.1
typing-extensions==4.14.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
websockets==14.2
xxhash==3.5.0
yarl==1.20.0
zstandard==0.23.0